refactor: keep integration filesystem wrappers in the CLI
Move the Node fs/process.env wrappers out of
@deepnote/database-integrations/node back into the CLI, leaving the package
as a single browser-safe entry point. The VS Code extension will provide its
own filesystem wrappers around the same browser-safe core.

- readDotEnv/updateDotEnv -> cli src/utils/dotenv.ts, using the dotenv
  library directly (drops the custom parseDotEnv reimplementation)
- resolveEnvVarRefs -> cli src/utils/env-var-refs.ts
- parseIntegrationsFile/getDefaultIntegrationsFilePath ->
  cli src/integrations/parse-integrations.ts
- readIntegrationsDocument/writeIntegrationsFile ->
  cli src/integrations/integrations-document.ts
- injectIntegrationEnvVars ->
  cli src/integrations/inject-integration-env-vars.ts

Remove the ./node export, the dotenv dependency, and the second tsdown
entry from @deepnote/database-integrations.
