+++ packages/cli/src/utils/dotenv.test.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs/promises'
 import os from 'node:os'
 import path from 'node:path'
 import { afterEach, beforeEach, describe, expect, it } from 'vitest'
-import { readDotEnv, updateDotEnv } from './index'
+import { readDotEnv, updateDotEnv } from './dotenv'
 
 describe('dotenv utilities', () => {
   let tempDir: string
